refactor: move cli src into separate subdirectory
